backend/requirements.txt
Django==3.2.18
djoser==2.1.0
django-environ==0.10.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
Pillow==9.4.0
drf-extra-fields==3.4.1